fix: project structure refactor
