+++ nextjs-logs-manager/src/components/DataTable/index.tsx
