+++ nextjs-logs-manager/src/components/Loader/index.tsx
