+++ nextjs-logs-manager/src/components/Navbar/index.tsx
