+++ nextjs-logs-manager/src/components/constants/columns.logs-view.tsx
