+++ nextjs-logs-manager/src/lib/authContext.tsx
