+++ nextjs-logs-manager/src/lib/generateHash.ts
