+++ nextjs-logs-manager/src/types/db.types.ts
