+++ nextjs-logs-manager/styles/globals.css
